pyproject.toml
[project]
name = "link-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.124.4",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "sqlmodel>=0.0.27",
]
